requirements/requirements.txt
mysql-connector-python==8.1.0
colorama==0.4.6
bcrypt==4.0.1
